forge fmt

+++ src/interfaces/IKintoApp.sol
@@ -8,21 +8,31 @@ interface IKintoApp {
         string name;
         address developerWallet; // the address that deploys the wallet
         bool dsaEnabled; // whether or not this application can request PII from users
-        uint rateLimitPeriod;
-        uint rateLimitNumber; // in txs
-        uint gasLimitPeriod;
-        uint gasLimitCost; // in eth
+        uint256 rateLimitPeriod;
+        uint256 rateLimitNumber; // in txs
+        uint256 gasLimitPeriod;
+        uint256 gasLimitCost; // in eth
     }
 
     /* ============ State Change ============ */
 
-    function registerApp(string calldata _name, address parentContract, address[] calldata childContracts, uint256[4] calldata appLimits) external;
+    function registerApp(
+        string calldata _name,
+        address parentContract,
+        address[] calldata childContracts,
+        uint256[4] calldata appLimits
+    ) external;
 
     function enableDSA(address app) external;
 
     function setSponsoredContracts(address _app, address[] calldata _contracts, bool[] calldata _flags) external;
 
-    function updateMetadata(string calldata _name, address parentContract, address[] calldata childContracts, uint256[4] calldata appLimits) external;
+    function updateMetadata(
+        string calldata _name,
+        address parentContract,
+        address[] calldata childContracts,
+        uint256[4] calldata appLimits
+    ) external;
 
     /* ============ Basic Viewers ============ */
 
@@ -31,7 +41,7 @@ interface IKintoApp {
     function symbol() external pure returns (string memory);
 
     function getContractLimits(address _contract) external view returns (uint256[4] memory);
-    
+
     function getAppMetadata(address _contract) external view returns (Metadata memory);
 
     function getContractSponsor(address _contract) external view returns (address);
@@ -43,5 +53,4 @@ interface IKintoApp {
     function DEVELOPER_ADMIN() external view returns (bytes32);
 
     function UPGRADER_ROLE() external view returns (bytes32);
-
 }
